requirements.txt
--extra-index-url https://dns-technologies.github.io/dbhose-dev-pip/simple/
apache-airflow>=2.4.3
native-dumper==0.3.7.dev3
pgpack-dumper==0.3.7.dev3
dr-herriot==0.1.0.dev0
